MAILET-115 Extract SpecialAddress from AbstractRedirect
