+++ server/mailet/mailets/src/main/java/org/apache/james/transport/mailets/redirect/SpecialAddressKind.java
@@ -1,0 +1,51 @@
+/****************************************************************
+ * Licensed to the Apache Software Foundation (ASF) under one   *
+ * or more contributor license agreements.  See the NOTICE file *
+ * distributed with this work for additional information        *
+ * regarding copyright ownership.  The ASF licenses this file   *
+ * to you under the Apache License, Version 2.0 (the            *
+ * "License"); you may not use this file except in compliance   *
+ * with the License.  You may obtain a copy of the License at   *
+ *                                                              *
+ *   http://www.apache.org/licenses/LICENSE-2.0                 *
+ *                                                              *
+ * Unless required by applicable law or agreed to in writing,   *
+ * software distributed under the License is distributed on an  *
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY       *
+ * KIND, either express or implied.  See the License for the    *
+ * specific language governing permissions and limitations      *
+ * under the License.                                           *
+ ****************************************************************/
+
+package org.apache.james.transport.mailets.redirect;
+
+public enum SpecialAddressKind {
+    SENDER("sender"),
+    REVERSE_PATH("reverse.path"),
+    FROM("from"),
+    REPLY_TO("reply.to"),
+    TO("to"),
+    RECIPIENTS("recipients"),
+    DELETE("delete"),
+    UNALTERED("unaltered"),
+    NULL("null");
+
+    private String value;
+
+    private SpecialAddressKind(String value) {
+        this.value = value;
+    }
+
+    public static SpecialAddressKind forValue(String value) {
+        for (SpecialAddressKind kind : values()) {
+            if (kind.value.equals(value)) {
+                return kind;
+            }
+        }
+        return null;
+    }
+
+    public String getValue() {
+        return value;
+    }
+}
